Added live purepursuit and live odom plot
Added the option to do live purepursuit based on the odom topic received from the turtlebot.

Added the option to plot the location recorded by odom. For live pure-pursuit this is done by default as the predicted and odom plot will be identical

Added functionality that resets odom topic when 'start' button is pushed